framework: Add fwk_str_is_in_boundry function
This change adds an additional function, fwk_str_is_in_boundry, that can
be used to check if a copy will be within the bounds of the destination
buffer.

This is a simple maths function, returns true if the copy will be in
bounds otherwise returns false.

Also adds associated tests.

Signed-off-by: Matthew Wellings <matthew.wellings@arm.com>